Items Wave 1 Batch A: 4 fortunes, Overcharge chance, Crystal Ball buff
New fortunes: La Partage (50% refund on 0 landings), The Bookkeeper
(+1 base on dozen/column wins), High Roller (+0.5 base on bets >=25%
of pre-spin balance), Rent Collector ($1 x stage per wheel segment
per round). New chance: Overcharge (select a segment, x3 multiplier
for one turn; MultiplierEffect factor generalized int->float).
Crystal Ball no longer lies (legendary buff). All registered in the
shop pools with icons. Design doc: .claude/design/new-items.md.

+++ core/src/main/java/de/mario/roguelette/wheel/effects/MultiplierEffect.java
@@ -1,9 +1,9 @@
 package de.mario.roguelette.wheel.effects;
 
 public class MultiplierEffect extends SegmentEffect implements MultiplierModifier {
-    private final int factor;
+    private final float factor;
 
-    public MultiplierEffect(int rounds, int factor) {
+    public MultiplierEffect(int rounds, float factor) {
         super(rounds);
         this.factor = factor;
     }
